Revert "Fix building with Xcode 12.5"
This reverts commit e926fb3642463c29956bb4daf032ca0d9772244c

+++ Source/OpenEmuHelperApp.swift
@@ -432,7 +432,7 @@ extension OSLog {
         
         if let parameters = parameters, let filter = _filterChain {
             for (key, val) in parameters {
-                filter.setValue(CGFloat(val), forParameterName: key)
+                filter.setValue(val, forParameterName: key)
             }
         }
     }
